Desktop: Fix startup error
